Organizadas as pastas e feita a página de detalhes
