+++ labedex/src/components/Header/styled.js
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const ContainerTitulo = styled.div`
+  height: 150px;
+  width: 100vw;
+  background-color: #f13c20;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  h1 {
+    position: relative;
+    left: 5vw;
+    font-size: 4em;
+    color: white;
+  }
+  img {
+    width: 130px;
+    margin-right: 5vw;
+  }
+  position: relative;
+`;
