+++ labedex/src/components/Screen/ScreenHome.js
@@ -5,7 +5,7 @@ const ContainerMain = styled.div`
   grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
   grid-gap: 15px;
   justify-self: center;
-  margin: 0 15px;
+  margin: 20px 15px;
 `;
 
 function Screen(props) {
